requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
playwright==1.50.0
pyee==12.1.1
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
